Fixed transaction bug
